# specs/001-newsletter-podcast-generator/contracts/api.yaml
openapi: 3.0.3
info:
  title: Newsletter to Podcast Generator API
  description: API for converting newsletter content into podcast episodes
  version: 1.0.0
  contact:
    name: Newsletter Podcast Generator
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.newsletter-podcast.com/v1
    description: Production server
paths:
  /newsletters:
    post:
      summary: Submit newsletter content for processing
      description: Create a new newsletter entry and start processing pipeline
      operationId: submit_newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsletterSubmission'
      responses:
        '201':
          description: Newsletter submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Newsletter'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Duplicate content (already processed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: List newsletters
      description: Retrieve list of submitted newsletters with pagination
      operationId: list_newsletters
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
      responses:
        '200':
          description: List of newsletters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsletterList'
...
  /newsletters/{newsletter_id}/retry:
    post:
      summary: Retry failed newsletter processing
      description: Restart processing for a failed newsletter
      operationId: retry_newsletter
      parameters:
        - name: newsletter_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Processing restart initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Newsletter'
        '400':
          description: Newsletter not in failed state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Newsletter not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    NewsletterSubmission:
      type: object
      required:
        - content
      properties:
        title:
          type: string
          maxLength: 500
          description: Newsletter title/subject
        url:
          type: string
          format: uri
          maxLength: 2048
          description: Source URL (if extracting from web)
        content:
          type: string
          minLength: 100
          maxLength: 50000
          description: Newsletter text content
        publication_date:
          type: string
          format: date-time
          description: Original publication date
    Newsletter:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        url:
          type: string
          format: uri
        content:
          type: string
        extracted_content:
          type: string
        publication_date:
          type: string
          format: date-time
        submitted_at:
          type: string
          format: date-time
        content_hash:
          type: string
        word_count:
          type: integer
        status:
          type: string
          enum: [pending, processing, completed, failed]
        episodes:
          type: array
          items:
            $ref: '#/components/schemas/Episode'
    NewsletterList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Newsletter'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    Episode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        newsletter_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        summary_text:
          type: string
        audio_url:
          type: string
          format: uri
        duration_seconds:
          type: integer
        file_size_bytes:
          type: integer
        publication_date:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [draft, generating_audio, uploading, published, failed]
        llm_service_used:
          type: string
        tts_service_used:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time